所有 between 验证完成

+++ src/main/java/com/tianyisoft/jvalidate/annotations/BetweenString.java
@@ -9,7 +9,7 @@ import java.lang.annotation.Target;
 @Target(ElementType.FIELD)
 @Retention(RetentionPolicy.RUNTIME)
 public @interface BetweenString {
-    String message() default "%s 长度必须在 %ld 和 %ld 之间";
-    long min();
-    long max();
+    String message() default "%s 长度必须在 %d 和 %d 之间";
+    int min();
+    int max();
 }
